+++ AgeOfEmpiresTaller/VistaSrc/CambioDeCoordendas.cpp
@@ -12,7 +12,7 @@ CambioDeCoordendas::CambioDeCoordendas(int ancho_por_celda,int alto_por_celda) {
 	ancho_celda=ancho_por_celda;
 	alto_celda=alto_por_celda;
 }
-void CambioDeCoordendas::transformar_isometrica_pantalla(float iso_x,float iso_y,int& pant_x,int& pant_y){
+void CambioDeCoordendas::transformar_isometrica_pantalla(double iso_x,double iso_y,int& pant_x,int& pant_y){
 	pant_x = (( iso_x - iso_y ) * this->ancho_por_celda())/2;
 	pant_y =(( iso_x + iso_y)*  this->altura_por_celda()) /2;
 }
@@ -21,16 +21,16 @@ void CambioDeCoordendas::transformar_isometrica_pantalla(int iso_x,int iso_y,int
 	pant_y =(( iso_x + iso_y)*  this->altura_por_celda()) /2;
 }
 
-void CambioDeCoordendas::transformar_pantalla_isometrica(int pant_x,int pant_y,float& iso_x,float& iso_y){
-	double termino_x=  ( (float)pant_x/ (this->ancho_por_celda()/2)) ;
-	double termino_y= ( (float)pant_y/ (this->altura_por_celda()/2) );
+void CambioDeCoordendas::transformar_pantalla_isometrica(int pant_x,int pant_y,double& iso_x,double& iso_y){
+	double termino_x=  ( (double)pant_x/ (this->ancho_por_celda()/2)) ;
+	double termino_y= ( (double)pant_y/ (this->altura_por_celda()/2) );
 	iso_x= (termino_x+ termino_y)/2;//MAGIA NEGRA
 	iso_y = (termino_y - termino_x)/2;//En serio no se porque funciona
 }
 
 void CambioDeCoordendas::transformar_pantalla_isometrica(int pant_x,int pant_y,int& iso_x,int& iso_y){
-	double termino_x=  ( (float)pant_x/ (this->ancho_por_celda()/2)) ;
-	double termino_y= ( (float)pant_y/ (this->altura_por_celda()/2) );
+	double termino_x=  ( (double)pant_x/ (this->ancho_por_celda()/2)) ;
+	double termino_y= ( (double)pant_y/ (this->altura_por_celda()/2) );
 	iso_x= (termino_x+ termino_y)/2;//MAGIA NEGRA
 	iso_y = (termino_y - termino_x)/2;//En serio no se porque funciona
 }
