+++ AgeOfEmpiresTaller/VistaSrc/CambioDeCoordendas.h
@@ -13,9 +13,9 @@ class CambioDeCoordendas {
 public:
 	CambioDeCoordendas(int ancho_celda,int alto_celda);
 	void transformar_pantalla_isometrica(int pant_x,int pant_y,int& iso_x,int& iso_y);
-	void transformar_pantalla_isometrica(int pant_x,int pant_y,float& iso_x,float& iso_y);
+	void transformar_pantalla_isometrica(int pant_x,int pant_y,double& iso_x,double& iso_y);
 	void transformar_isometrica_pantalla(int iso_x,int iso_y,int& pant_x,int& pant_y);
-	void transformar_isometrica_pantalla(float iso_x,float iso_y,int& pant_x,int& pant_y);
+	void transformar_isometrica_pantalla(double iso_x,double iso_y,int& pant_x,int& pant_y);
 	virtual ~CambioDeCoordendas();
 private:
 	int ancho_por_celda();
